# doc/api.yaml
openapi: 3.0.1
info:
  title: WasaPhoto
  description: |
    API specification for WASAPhoto, University project
    for web and software architecture course.
  version: 1.0.2
servers:
  - url: http://localhost:3000
...
paths:
...
  /users/{userId}/followers/{followerId}:
    parameters:
      - $ref: "#/components/parameters/userId" #followed by followerId
      - $ref: "#/components/parameters/followerId" #followed

    put:
      security:
        - bearerAuth: []
      tags:
        - Follow
      summary: follow a user
      
      description: the action of following a user y from user x
      operationId: followUser

      responses:
        "204":
          description: user followed succesfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      security:
        - bearerAuth: []
      tags:
        - Follow
      summary: unfollow a user
      description: the action of unfollowin a user y from user x
      operationId: unfollowUser
      responses:
        "204":
          description: follower removed succesfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/{userId}/bans/{bannerId}:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/bannerId"

    put:
      security:
        - bearerAuth: []
      tags:
        - Ban
      summary: ban a user
      description: the action of banning a user y from user x return username of banned user
      operationId: banUser

      responses:
        "204":
          description: user banned succesfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      security:
        - bearerAuth: []
      tags:
        - Ban
      summary: unban a user
      description: the action of removing a ban status from a y user by user x
      operationId: unbanUser

      responses:
        "204":
          description: user unbanned succesfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"

  #return here
  /users/{userId}/photos:
    parameters:
      - $ref: "#/components/parameters/userId"
    post:
      security:
        - bearerAuth: []
      tags:
        - Photo
      summary: upload of a photo
      description: |-
        posts a new photo on the user's profile, returns the photoid
      operationId: uploadPhoto
      requestBody:
        description: the file of the photo to be uploaded
        content:
          image/png:
            schema:
              $ref: "#/components/schemas/PhotoFile"
          image/jpeg:
            schema:
              $ref: "#/components/schemas/PhotoFile"
        required: true

      responses:
        "200":
          description: photo uploaded succesfully
          content: 
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
        
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/{userId}/photos/{photoId}:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/photoId"

    get:
      security:
        - bearerAuth: []
      tags:
        - Photo
      summary: retrieves a photo
      description: returns an Image
      operationId: servePhoto
      responses:
        "200":
          description: photo retrieved succesfully
          content:
            image/png:
              schema:
                $ref: "#/components/schemas/PhotoFile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      security:
        - bearerAuth: []
      tags:
        - Photo
      summary: delete a photo
      description: removes a photo from a user profile
      operationId: deletePhoto
      responses:
        "204":
          description: photo removed successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/{userId}/photos/{photoId}/comments:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/photoId"

    post:
      security:
        - bearerAuth: []
      tags:
        - Comments
      summary: comments a photo
      description: leaves a comment under a photo
      operationId: commentPhoto
      requestBody:
        description: the body of the comment
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentBody"

      responses:
        "201":
          description: comment left succesfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
                
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/{userId}/photos/{photoId}/comments/{commentId}:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/photoId"
      - $ref: "#/components/parameters/commentId"

    delete:
      security:
        - bearerAuth: []
      tags:
        - Comments
      summary: uncomments a photo
      description: removes a comment from a photo
      operationId: uncommentPhoto
      responses:
        "204":
          description: comment removed successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/{userId}/photos/{photoId}/likes/{likerId}:
    parameters:
      - $ref: "#/components/parameters/photoId"
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/likerId"

    put:
      security:
        - bearerAuth: []
      tags:
        - Like
      summary: adds a like to a photo
      description: the action of putting a like to a photo from the feed; userId will be the identifier for a like in that context
      operationId: likePhoto
      responses:
        "204":
          description: photo liked successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      security:
        - bearerAuth: []
      tags:
        - Like
      summary: removes a like from a photo
      description: the action of removing a like from a photo from the feed
      operationId: unlikePhoto
      responses:
        "204":
          description: photo unliked successfully

        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
...
    BadRequest:
      description: syntax of the request was not compliant with docs, check parameters
    #403
    Forbidden:
      description: server will not accept request due to lacking of client permissions
    #404
    NotFound:
      description: resource not found

    #500 class responses
    #500
    InternalServerError:
      description: the server encountered an error while processing the request
  parameters:
    userId:
      description: identifier for a single user
      schema:
        $ref: "#/components/schemas/Identifier"
      name: userId
      in: path
      required: true
    followerId:
      description: identifier for a user that follows another
      schema:
        $ref: "#/components/schemas/Identifier"
      name: followerId
      in: path
      required: true
    bannerId:
      description: identifier for a user banning another
      schema:
        $ref: "#/components/schemas/Identifier"
      name: bannerId
      in: path
      required: true
    photoId:
      description: identifier for a photo
      schema:
        $ref: "#/components/schemas/Identifier"
      name: photoId
      in: path
      required: true
    commentId:
      description: identifier for a comment, used to retrieve comments
      schema:
        $ref: "#/components/schemas/Identifier"
      name: commentId
      in: path
      required: true
    likerId:
      name: likerId
      in: path
      required: true
      description: identifier for a user that likes a photo (it represents the actual like)
      schema:
        $ref: "#/components/schemas/Identifier"
  schemas:
    Identifier:
      title: identifier
      description: alphanumerical identifier for a resource
      type: string
      readOnly: true
      #consider specializing it
      pattern: "^@[A-Z][0-9]{9}$"
      example: "@U000000001"
      minLength: 11
      maxLength: 11
    UserName:
      title: username
      description: user name associated with userId, part of a user obj
      type: string
      example: neosnakex34
      pattern: "^[a-z0-9]*?$"
      minLength: 3
      maxLength: 12
...
    PhotoFile:
      title: photoFile
      description: the actual file of the photo, as a stream of bytes retrieved by it's path in filesystem
      properties:
        path:
          description: position in filesystem for a file
          type: string
          example: "/tmp/wasaphoto/photofiles/rei.jpg"

        file:
          description: the actual file of the photo
          type: string
          format: binary
          minLength: 8 # to fit check for pngs header
          maxLength: 10485760 #binary 10 megs (2 power of 20 times 10)
    Photo:
      title: photo
      description: a photo on the platform
      type: object
      properties:
        photoId:
          $ref: "#/components/schemas/Identifier"

        # uploader user id
        uploaderUserId:
          $ref: "#/components/schemas/Identifier"
        
        uploaderUsername:
          $ref: "#/components/schemas/UserName"
        
        likes:
          description: number of likes the photo has recieved (represented by userId that liked the photo)
          type: integer
          example: 50

        likedByLoggedUser:
          description: boolean value that tells if the photo is liked by the logged user will be computed when retrieving photo
          type: boolean
          example: true

        comments:
          description: comments attached to the photo, made by users
          type: array
          minItems: 0
          maxItems: 20 # not for real use, just as an example
          items:
            $ref: "#/components/schemas/Comment"

        date:
          description: date of uploading
          type: string
          format: date-time
          example: "2021-05-12T09:00:00Z"
          minLength: 20
          maxLength: 20

        # what will be a photopath in struct (path + data)
        rawPhotoFile:
          $ref: "#/components/schemas/PhotoFile"
    CommentBody:
      description: body of the comment
      type: string
      minLength: 1
      maxLength: 400
      pattern: "^.*?$"
      example: "nice one bro"
    Comment:
      title: comment
      description: a comment on a photo
      type: object
      properties:
        commentId:
          $ref: "#/components/schemas/Identifier"

        userId:
          $ref: "#/components/schemas/Identifier"

        photoId:
          $ref: "#/components/schemas/Identifier"

        body:
          $ref: "#/components/schemas/CommentBody"

        date:
          description: date of the comment
          type: string
          format: date-time
          example: "2023-12-12T09:00:00Z"
          minLength: 20
          maxLength: 20